20260225_workout_daily/requirements.txt
garminconnect>=0.2.38
anthropic
python-dotenv
